introduce data providers
data providers introduce a new way of adding additional data to the
events.
Previously it was static and one could not decide which data should be
included. Adding new data always required a change to the plugin itself.
In line mode we sent all the data that is more or less static with each
event. We could now only sent the core data and introduce a final event
that sends all the other data at the end.